Make the bib merger more resource agnostic
