--- sierra_adapter/sierra_bib_merger/docker-compose.yml
@@ -1,8 +1,0 @@
-sqs:
- image: s12v/elasticmq
- ports:
-   - "9324:9324"
-dynamodb:
-  image: peopleperhour/dynamodb
-  ports:
-    - "45678:8000"
