+++ sierra_adapter/sierra_merger/src/main/resources/logback.xml
