+++ sierra_adapter/sierra_merger/src/main/scala/uk/ac/wellcome/platform/sierra_merger/controllers/ManagementController.scala
@@ -1,0 +1,13 @@
+package uk.ac.wellcome.platform.sierra_merger.controllers
+
+import javax.inject.{Inject, Singleton}
+
+import com.twitter.finagle.http.Request
+import com.twitter.finatra.http.Controller
+
+@Singleton
+class ManagementController @Inject()() extends Controller {
+  get("/management/healthcheck") { request: Request =>
+    response.ok.json(Map("message" -> "ok"))
+  }
+}
