+++ sierra_adapter/sierra_merger/src/test/resources/logback.xml
